requirements.txt
python-telegram-bot[webhooks]==22.3
spotipy>=2.24.0
redis>=5.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
